+++ CHANGELOG.md
@@ -26,8 +26,7 @@
 ### Fixed
 - Fix publish-findings forEach try/catch dropping rest of batch on first error [(#49)](https://github.com/wazuh/wazuh-indexer-alerting/pull/49)
 - Fix SLF4J startup warning by adding Log4j2 provider [(#74)](https://github.com/wazuh/wazuh-indexer-alerting/pull/74)
-
-  
+- Fix memory leak in doc-level monitor fan-out [(#106)](https://github.com/wazuh/wazuh-indexer-alerting/pull/106)
 
 ### Security